fix(parsers): beta is a float not int

+++ src/parsers.cpp
@@ -14,7 +14,7 @@ BedLine parseBEDRecord(const std::string& bedString) {
     std::string chrom = fields[0];
     int start = std::stoi(fields[1]);
     int end = std::stoi(fields[2]);
-    int beta = std::stoi(fields[3]);
+    float beta = std::stof(fields[3]);
     int cov = std::stoi(fields[4]);
 
     BedLine read = {chrom, start, end, beta, cov};
